refactor: login response docs
