+++ packages/server/src/controllers/session/register.ts
